Added all panels to chose cylinder for truck. Things to do: logic in model and controller and add all data to db
